# Speed up build and dev by getting rid of Babel