--- a/apps/saleor-app-checkout/.eslintrc.js
+++ b/apps/saleor-app-checkout/.eslintrc.js
@@ -3,4 +3,12 @@ module.exports = {
   parserOptions: {
     tsconfigRootDir: __dirname,
   },
+  rules: {
+    "formatjs/enforce-id": [
+      "error",
+      {
+        idInterpolationPattern: "[folder]/[name]/[sha512:contenthash:base64:6]",
+      },
+    ],
+  },
 };