--- a/apps/saleor-app-checkout/__tests__/backend/payments/providers/adyen/middlewares.test.ts
+++ b/apps/saleor-app-checkout/__tests__/backend/payments/providers/adyen/middlewares.test.ts
@@ -61,7 +61,7 @@ const requestWithNotification = {
   },
 };
 
-const mockPrivateSettings = (returnMockedSettings: boolean = true) => {
+const mockPrivateSettings = (returnMockedSettings = true) => {
   if (!returnMockedSettings) {
     mockedGetPrivateSettings.mockResolvedValueOnce({
       paymentProviders: {