--- a/apps/saleor-app-checkout/content/locales/en.json
+++ b/apps/saleor-app-checkout/content/locales/en.json
@@ -1,1 +1,214 @@
-{}
+{
+  "ChannelDetails/messages/1+HS9r": {
+    "developer_comment": "section title",
+    "string": "Select payment methods"
+  },
+  "ChannelList/messages/vOL1vk": {
+    "developer_comment": "label",
+    "string": "Channel name"
+  },
+  "CustomizationDetails/messages/CAoP49": {
+    "developer_comment": "section title",
+    "string": "Customization preview"
+  },
+  "CustomizationDetails/messages/Z9ndZ7": {
+    "developer_comment": "information",
+    "string": "Please provide valid checkout URL"
+  },
+  "CustomizationDetails/messages/eQTUqD": {
+    "developer_comment": "input label",
+    "string": "Checkout URL"
+  },
+  "ErrorDetails/messages/+1z3pp": {
+    "developer_comment": "error title",
+    "string": "Something went wrong"
+  },
+  "FileInput/messages/AlmJPi": {
+    "developer_comment": "button",
+    "string": "Upload file"
+  },
+  "FileInput/messages/Zr2A5Y": {
+    "developer_comment": "description",
+    "string": "Max size 1GB"
+  },
+  "FileInput/messages/h6LHdF": {
+    "developer_comment": "description",
+    "string": "Drag an image here"
+  },
+  "PaymentProviderDetails/messages//1yAUF": {
+    "developer_comment": "notice for private settings",
+    "string": "These information will not be encrypted"
+  },
+  "PaymentProviderDetails/messages/OOy0zB": {
+    "developer_comment": "notice for private settings",
+    "string": "These information will be encrypted"
+  },
+  "PaymentProviderDetails/messages/n3PNHM": {
+    "developer_comment": "section title",
+    "string": "Payment provider settings"
+  },
+  "messages/customization//dYclm": {
+    "developer_comment": "customization setting",
+    "string": "Button BG Hover"
+  },
+  "messages/customization/0NYFC/": {
+    "developer_comment": "customization setting",
+    "string": "Button Text"
+  },
+  "messages/customization/4kEXS+": {
+    "developer_comment": "customization",
+    "string": "Branding"
+  },
+  "messages/customization/D8irjQ": {
+    "developer_comment": "customization setting",
+    "string": "Button BG Primary"
+  },
+  "messages/customization/DopY6P": {
+    "developer_comment": "customization setting",
+    "string": "Error"
+  },
+  "messages/customization/HpJ0xA": {
+    "developer_comment": "customization setting",
+    "string": "Logo"
+  },
+  "messages/customization/S/xcUp": {
+    "developer_comment": "customization setting",
+    "string": "Success"
+  },
+  "messages/customization/X5y+L2": {
+    "developer_comment": "customization",
+    "string": "Product settings"
+  },
+  "messages/customization/cNpy7H": {
+    "developer_comment": "customization setting",
+    "string": "Text"
+  },
+  "messages/customization/e9Fc8u": {
+    "developer_comment": "customization setting",
+    "string": "Low stock threshold"
+  },
+  "messages/customization/m/r08F": {
+    "developer_comment": "customization setting",
+    "string": "Border Primary"
+  },
+  "messages/paymentMethods/6bexF4": {
+    "developer_comment": "payment method",
+    "string": "Drop-in"
+  },
+  "messages/paymentMethods/BFIFWS": {
+    "developer_comment": "payment method",
+    "string": "Apple Pay"
+  },
+  "messages/paymentMethods/IKIBYk": {
+    "developer_comment": "payment method",
+    "string": "Credit card"
+  },
+  "messages/paymentMethods/nXqPc7": {
+    "developer_comment": "payment method",
+    "string": "PayPal"
+  },
+  "messages/paymentProviders/+1my3D": {
+    "developer_comment": "payment provider setting",
+    "string": "Webhook HMAC key"
+  },
+  "messages/paymentProviders/9hBUgM": {
+    "developer_comment": "payment provider setting",
+    "string": "API key"
+  },
+  "messages/paymentProviders/CH2iz9": {
+    "developer_comment": "payment provider setting",
+    "string": "Webhook password"
+  },
+  "messages/paymentProviders/CQH/Vb": {
+    "developer_comment": "payment provider",
+    "string": "Mollie"
+  },
+  "messages/paymentProviders/I6I7xw": {
+    "developer_comment": "payment provider setting",
+    "string": "Webhook username"
+  },
+  "messages/paymentProviders/WdX1+f": {
+    "developer_comment": "payment provider setting",
+    "string": "Public client key"
+  },
+  "messages/paymentProviders/YBcvvL": {
+    "developer_comment": "payment provider setting",
+    "string": "Publishable Key"
+  },
+  "messages/paymentProviders/ehp06H": {
+    "developer_comment": "payment provider",
+    "string": "Stripe"
+  },
+  "messages/paymentProviders/fQ20+Q": {
+    "developer_comment": "payment provider setting",
+    "string": "Secret Key"
+  },
+  "messages/paymentProviders/glq8nA": {
+    "developer_comment": "payment provider setting",
+    "string": "Webhookd Secret"
+  },
+  "messages/paymentProviders/m5P19o": {
+    "developer_comment": "payment provider",
+    "string": "Adyen"
+  },
+  "messages/paymentProviders/qHRtb0": {
+    "developer_comment": "payment provider setting",
+    "string": "Private API key"
+  },
+  "messages/paymentProviders/t7YEMr": {
+    "developer_comment": "payment provider setting",
+    "string": "Profile ID"
+  },
+  "messages/paymentProviders/ufY8+j": {
+    "developer_comment": "payment provider setting",
+    "string": "Merchant account"
+  },
+  "misc/commonMessages/95qy1d": {
+    "developer_comment": "section name",
+    "string": "Channels"
+  },
+  "misc/commonMessages/g8hTTQ": {
+    "developer_comment": "section name",
+    "string": "Settings"
+  },
+  "misc/commonMessages/gnUZny": {
+    "developer_comment": "section name",
+    "string": "Customization"
+  },
+  "misc/errorMessages/KEl0/z": {
+    "developer_comment": "error message",
+    "string": "Something went wrong"
+  },
+  "misc/errorMessages/OvAW89": {
+    "developer_comment": "error message",
+    "string": "Payment provider not found."
+  },
+  "misc/errorMessages/fHPUmP": {
+    "developer_comment": "error message",
+    "string": "Unknown error"
+  },
+  "misc/errorMessages/ftpe3L": {
+    "developer_comment": "error message",
+    "string": "Invalid"
+  },
+  "misc/errorMessages/nE/aC7": {
+    "developer_comment": "error message",
+    "string": "Channel payment options not found."
+  },
+  "misc/errorMessages/ofWxwJ": {
+    "developer_comment": "error message",
+    "string": "Required"
+  },
+  "misc/errorMessages/v74OtS": {
+    "developer_comment": "error message",
+    "string": "Not updated"
+  },
+  "misc/errorMessages/vZ8Jjv": {
+    "developer_comment": "error message",
+    "string": "GraphQL error"
+  },
+  "misc/errorMessages/z3iWGt": {
+    "developer_comment": "error message",
+    "string": "Not found"
+  }
+}