--- a/apps/saleor-app-checkout/frontend/components/elements/FileInput/messages.ts
+++ b/apps/saleor-app-checkout/frontend/components/elements/FileInput/messages.ts
@@ -3,14 +3,17 @@ import { defineMessages } from "react-intl";
 export const messages = defineMessages({
   dragImage: {
     defaultMessage: "Drag an image here",
+    id: "FileInput/messages/h6LHdF",
     description: "description",
   },
   maxFileSize: {
     defaultMessage: "Max size 1GB",
+    id: "FileInput/messages/Zr2A5Y",
     description: "description",
   },
   uploadFile: {
     defaultMessage: "Upload file",
+    id: "FileInput/messages/AlmJPi",
     description: "button",
   },
 });