--- a/apps/saleor-app-checkout/frontend/components/templates/ChannelDetails/messages.ts
+++ b/apps/saleor-app-checkout/frontend/components/templates/ChannelDetails/messages.ts
@@ -3,6 +3,7 @@ import { defineMessages } from "react-intl";
 export const messages = defineMessages({
   selectPaymentMethods: {
     defaultMessage: "Select payment methods",
+    id: "ChannelDetails/messages/1+HS9r",
     description: "section title",
   },
 });