--- a/apps/saleor-app-checkout/frontend/components/templates/ChannelList/messages.ts
+++ b/apps/saleor-app-checkout/frontend/components/templates/ChannelList/messages.ts
@@ -3,6 +3,7 @@ import { defineMessages } from "react-intl";
 export const messages = defineMessages({
   channelName: {
     defaultMessage: "Channel name",
+    id: "ChannelList/messages/vOL1vk",
     description: "label",
   },
 });