--- a/apps/saleor-app-checkout/frontend/components/templates/CustomizationDetails/messages.ts
+++ b/apps/saleor-app-checkout/frontend/components/templates/CustomizationDetails/messages.ts
@@ -3,14 +3,17 @@ import { defineMessages } from "react-intl";
 export const messages = defineMessages({
   customizationPreview: {
     defaultMessage: "Customization preview",
+    id: "CustomizationDetails/messages/CAoP49",
     description: "section title",
   },
   checkoutUrl: {
     defaultMessage: "Checkout URL",
+    id: "CustomizationDetails/messages/eQTUqD",
     description: "input label",
   },
   noCheckoutUrl: {
     defaultMessage: "Please provide valid checkout URL",
+    id: "CustomizationDetails/messages/Z9ndZ7",
     description: "information",
   },
 });