--- a/apps/saleor-app-checkout/frontend/components/templates/ErrorDetails/messages.ts
+++ b/apps/saleor-app-checkout/frontend/components/templates/ErrorDetails/messages.ts
@@ -3,6 +3,7 @@ import { defineMessages } from "react-intl";
 export const messages = defineMessages({
   somethingWentWrong: {
     defaultMessage: "Something went wrong",
+    id: "ErrorDetails/messages/+1z3pp",
     description: "error title",
   },
 });