--- a/apps/saleor-app-checkout/frontend/components/templates/PaymentProviderDetails/messages.ts
+++ b/apps/saleor-app-checkout/frontend/components/templates/PaymentProviderDetails/messages.ts
@@ -3,14 +3,17 @@ import { defineMessages } from "react-intl";
 export const messages = defineMessages({
   paymentProviderSettings: {
     defaultMessage: "Payment provider settings",
+    id: "PaymentProviderDetails/messages/n3PNHM",
     description: "section title",
   },
   encryptedSettingNotice: {
     defaultMessage: "These information will be encrypted",
+    id: "PaymentProviderDetails/messages/OOy0zB",
     description: "notice for private settings",
   },
   publicSettingNotice: {
     defaultMessage: "These information will not be encrypted",
+    id: "PaymentProviderDetails/messages//1yAUF",
     description: "notice for private settings",
   },
 });