--- a/apps/saleor-app-checkout/frontend/misc/errorMessages.ts
+++ b/apps/saleor-app-checkout/frontend/misc/errorMessages.ts
@@ -3,34 +3,42 @@ import { defineMessages } from "react-intl";
 export const commonErrorMessages = defineMessages({
   somethingWentWrong: {
     defaultMessage: "Something went wrong",
+    id: "misc/errorMessages/KEl0/z",
     description: "error message",
   },
   unknownError: {
     defaultMessage: "Unknown error",
+    id: "misc/errorMessages/fHPUmP",
     description: "error message",
   },
   graphqlError: {
     defaultMessage: "GraphQL error",
+    id: "misc/errorMessages/vZ8Jjv",
     description: "error message",
   },
   unknown: {
     defaultMessage: "Unknown error",
+    id: "misc/errorMessages/fHPUmP",
     description: "error message",
   },
   invalid: {
     defaultMessage: "Invalid",
+    id: "misc/errorMessages/ftpe3L",
     description: "error message",
   },
   required: {
     defaultMessage: "Required",
+    id: "misc/errorMessages/ofWxwJ",
     description: "error message",
   },
   notFound: {
     defaultMessage: "Not found",
+    id: "misc/errorMessages/z3iWGt",
     description: "error message",
   },
   notUpdated: {
     defaultMessage: "Not updated",
+    id: "misc/errorMessages/v74OtS",
     description: "error message",
   },
 });
@@ -38,10 +46,12 @@ export const commonErrorMessages = defineMessages({
 export const notFoundMessages = defineMessages({
   paymentProviderNotFound: {
     defaultMessage: "Payment provider not found.",
+    id: "misc/errorMessages/OvAW89",
     description: "error message",
   },
   channelPaymentOptionsNotFound: {
     defaultMessage: "Channel payment options not found.",
+    id: "misc/errorMessages/nE/aC7",
     description: "error message",
   },
 });