--- a/apps/saleor-app-checkout/package.json
+++ b/apps/saleor-app-checkout/package.json
@@ -12,7 +12,7 @@
     "lint:staged": "next lint --dir . --fix",
     "check-types": "tsc --noEmit",
     "generate": "DOTENV_CONFIG_PATH=.env graphql-codegen -r env-vars --config codegen.yml",
-    "extract:i18n": "formatjs extract '{pages,components,hooks,misc}/**/*.{js,ts,tsx}' --format transifex --id-interpolation-pattern '[folder]/[name]/[sha512:contenthash:base64:6]' --out-file content/locales/en.json",
+    "extract:i18n": "formatjs extract '{backend,config,content,frontend,graphql,pages,types,utils}/**/*.{js,jsx,ts,tsx}' --format transifex --id-interpolation-pattern '[folder]/[name]/[sha512:contenthash:base64:6]' --out-file content/locales/en.json",
     "compile:i18n": "formatjs compile-folder --ast --format transifex content/locales content/compiled-locales",
     "i18n": "npm run extract:i18n && npm run compile:i18n",
     "clean": "rm -rf .next .turbo"