--- a/apps/saleor-app-checkout/setupTestsAfterEnv.ts
+++ b/apps/saleor-app-checkout/setupTestsAfterEnv.ts
@@ -18,6 +18,7 @@ afterEach(() => server.resetHandlers());
 afterEach(() => {
   consoleTypes.forEach((type) => {
     // @ts-expect-error
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-call
     typeof console[type]?.mockClear === "function" && console[type].mockClear();
   });
 });