--- a/apps/saleor-app-checkout/setupTestsBeforeEnv.ts
+++ b/apps/saleor-app-checkout/setupTestsBeforeEnv.ts
@@ -6,6 +6,7 @@ process.env.SALEOR_APP_TOKEN = "TEST";
 // Fix for jose missing TextEncoder error
 // https://github.com/inrupt/solid-client-authn-js/issues/1676
 if (typeof global.TextEncoder === "undefined") {
+  // eslint-disable-next-line @typescript-eslint/no-var-requires
   const { TextEncoder, TextDecoder } = require("util");
   global.TextEncoder = TextEncoder;
   global.TextDecoder = TextDecoder;