… apps/saleor-app-checkout/.babelrc
… apps/storefront/package.json
… pnpm-lock.yaml