- [x] get rid of babel
- [x] replace formatjs babel plugin with eslint plugin
- [x] fix eslint configuration to lint all files

|              | **before** | **after** |  **speedup**  | 
|--------------|:----------:|:---------:|:------------|
|    **cache** |   24.69s   |   22.82s  |  ✅ **–7.57%**  | 
| **no-cache** |   80.37s   |   59.59s  | ✅ **–25.86%** | 

Fixes #479